requirements.txt

Flask==2.1.1
discord.py==1.7.3
requests==2.26.0
flask-cors==3.0.10
werkzeug==2.0.1
websockets
aiosqlite
asyncio
flask-socketio